% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sample_score.R
\name{compute_npscore}
\alias{compute_npscore}
\title{compute nscore or pscore matrix}
\usage{
compute_npscore(ccs, w1matrix1, nscore = FALSE)
}
\arguments{
\item{nscore}{a logical value for switch evaluate the 
\code{pscore} or \code{nscore}.}
}
\description{
compute nscore or pscore matrix
}
