% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher.R
\name{.enrich}
\alias{.enrich}
\title{Fisher enrichment test}
\usage{
.enrich(cluster, geneSet, N)
}
\arguments{
\item{N}{the background size, number of unique gene id in
the background model.}
}
\description{
Fisher enrichment test
}
