% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computegenelist.R
\name{pfsnet.computegenelist}
\alias{pfsnet.computegenelist}
\title{compute gene list}
\usage{
pfsnet.computegenelist(w, beta)
}
\value{
a list of gene id which theirs expression is 
greater than threshold \code{beta}.
}
\description{
compute the gene list which its 
members expression level is greater than \code{beta}.
}
