% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher.R
\name{sizeof}
\alias{sizeof}
\title{Get background size}
\usage{
sizeof(background)
}
\arguments{
\item{background}{a model object which is created from
the \code{\link{ReadFisherCluster}} function.}
}
\description{
Get background size
}
